requirements/requirements.txt
torch

numpy<2
pandas
matplotlib
seaborn

py-cpuinfo

nguyenpanda
tqdm
typing-extensions
